// models/AddFestiveModal.tsx
export interface FestiveData {
  bannerContentId?: string;
  endDate?:string;
  endTime?:string;
  entityId?: string;
  failContentId?:string;
  failMessage?:string;
  id?: string;
  maxAmount?: string;
  minAmount?: string;
  packageDescription?: string;
  seasonCode?: string;
  seasonDescription?:string;
  seasonNumber?: string;
  startDate?: string;
  startTime?:string;
  statusCode?: string;
  successContentId?: string;
  successMessage?:string;
  summaryContentId?:string;
}